Improved the thread system and added Im Gui backend client.
* The threading library is heavily improved. Threads are managed by the systems itself and is accessible from anywhere of the engine and client application via the system locator service.
* Added support for the Im Gui backend client.

+++ Source/Runtime/VulkanRBL/Primitives/VulkanTexture.cpp
@@ -63,11 +63,17 @@ namespace Dynamik
 
 		void VulkanTexture::terminate(RCoreObject* pCoreObject)
 		{
-			pImage->terminate(pCoreObject);
-			StaticAllocator<VulkanImage>::rawDeallocate(pImage);
+			if (pImage)
+			{
+				pImage->terminate(pCoreObject);
+				StaticAllocator<VulkanImage>::rawDeallocate(pImage);
+			}
 
-			pSampler->terminate(pCoreObject);
-			StaticAllocator<VulkanImageSampler>::rawDeallocate(pSampler);
+			if (pSampler)
+			{
+				pSampler->terminate(pCoreObject);
+				StaticAllocator<VulkanImageSampler>::rawDeallocate(pSampler);
+			}
 		}
 
 		VulkanTexture::operator VulkanImage() const
